refactor: blocks test folder
